chore: refactor posthog
