+++ apps/api/package.json
@@ -19,7 +19,6 @@
     "express": ">=5.0.0-beta.1",
     "express-openapi-validator": "^5.0.2",
     "pino-http": "^8.3.3",
-    "posthog-node": "^2.5.3",
     "simple-oauth2": "^5.0.0",
     "snakecase-keys": "^5.4.5",
     "tslib": "^2.5.0",
