+++ apps/api/services/sync_service.ts
@@ -1,4 +1,3 @@
-import { distinctId } from '@/lib/distinct_identifier';
 import { CRM_COMMON_MODELS } from '@supaglue/core/types/';
 import { CommonModel } from '@supaglue/core/types/common';
 import { SyncInfo } from '@supaglue/core/types/sync_info';
@@ -55,7 +54,7 @@ export class SyncService {
           workflowType: runSyncs,
           workflowId: getRunSyncsWorkflowId(connectionId),
           taskQueue: SYNC_TASK_QUEUE,
-          args: [{ connectionId, sessionId: distinctId }],
+          args: [{ connectionId }],
         },
         state: {
           triggerImmediately: true,
