+++ packages/core/package.json
@@ -17,6 +17,7 @@
     "pg-copy-streams": "^6.0.5",
     "pino": "^8.11.0",
     "pino-pretty": "^9.4.0",
+    "posthog-node": "^2.5.4",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
